update for buildroot

+++ amsubdec/amsub_codec/libsub_idxsub/amsub_idxsub_dec.h
@@ -89,7 +89,7 @@ typedef struct
     int height;
     unsigned short colorcode;
     unsigned short  contrast;
-    unsigned long prtData;
+    unsigned prtData;
     unsigned cls;
 } Vob_subtitle_showdata;
